Always include a space in a transform matrix.
As I noted in: https://github.com/RazrFalcon/SVGCleaner/pull/28 my previous patch seems to have made it so that a condition is always matched when checking to add spaces in between matrix points. I can't think of a time when one of those conditions isn't matched, so this cleans up the code.